Move also the ArgocdService implementation to the -node package
... to avoid forcing users to redefine the serviceRef or have circular dependencies
between the -node and -backend packages
