+++ plugins/backend/backstage-plugin-argo-cd-backend/src/service/timer.services.ts
@@ -1,3 +1,3 @@
-export function timer(ms: number) {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
+// timer has moved to @roadiehq/backstage-plugin-argo-cd-node.
+// Re-exported here for backwards compatibility.
+export { timer } from '@roadiehq/backstage-plugin-argo-cd-node';
