+++ plugins/backend/backstage-plugin-argo-cd-backend/src/utils/getArgoConfig.ts
@@ -1,14 +1,2 @@
-import { InstanceConfig } from '../service/types';
-
-export const getArgoConfigByInstanceName = ({
-  argoConfigs,
-  argoInstanceName,
-}: {
-  argoConfigs: InstanceConfig[];
-  argoInstanceName: string;
-}) => {
-  const matchedArgoConfig = argoConfigs.find(
-    configs => configs.name === argoInstanceName,
-  );
-  return matchedArgoConfig;
-};
+// Moved to @roadiehq/backstage-plugin-argo-cd-node. Re-exported for backwards compatibility.
+export { getArgoConfigByInstanceName } from '@roadiehq/backstage-plugin-argo-cd-node';
