+++ plugins/backend/backstage-plugin-argo-cd-node/package.json
@@ -49,7 +49,8 @@
   },
   "dependencies": {
     "@backstage/backend-plugin-api": "backstage:^",
-    "@backstage/config": "backstage:^"
+    "@backstage/config": "backstage:^",
+    "cross-fetch": "^3.1.4"
   },
   "devDependencies": {
     "@backstage/cli": "backstage:^"
